Put data retrieval in separate file
